minimo de logica dentro dos componentes de carta
