+++ src/app/components/cartas/carta/carta.component.css
