+++ src/app/components/jogo/criar-jogo/criar-jogo.component.html
@@ -1,16 +1,17 @@
-<button class="btn btn-primary btn-sm" *ngIf="!criandoJogo" (click)="criarJogo()">+</button>
+<button class="btn btn-primary btn-sm" *ngIf="!criandoJogo" (click)="iniciar()">+</button>
 <div class="col-md-6" *ngIf="criandoJogo">
   <form #f="ngForm">
     <div class="form-group">
-      <input [(ngModel)]="jogoNome" name="jogoNome" type="text" class="form-control" placeholder="Jogo Nome" required>
+      <input [(ngModel)]="jogoNome" name="jogoNome" type="text" class="form-control" placeholder="Nome do jogo"
+        required>
     </div>
     <div class="form-group">
-      <button class="btn btn-primary" (click)="criarJogo()">
+      <button class="btn btn-primary" (click)="fechar()">
         Cancelar
       </button>
-      <button class="btn btn-primary" [disabled]="!f.valid" (click)="salvarJogo()">
+      <button class="btn btn-primary" [disabled]="!f.valid" (click)="salvar()">
         Criar
-      </button>        
+      </button>
     </div>
   </form>
 </div>
