+++ src/app/service/jogo.service.ts
@@ -108,19 +108,6 @@ export class JogoService {
         return jogadoresParticipantes.length < 1;
     }
 
-    getVisaoCarta(): boolean {
-        var visaoCarta = this.cookieService.get("visaoCarta");
-        if (visaoCarta === undefined || visaoCarta === null || visaoCarta === "") {
-            visaoCarta = "true";
-            this.setVisaoCarta(visaoCarta)
-        }
-        return visaoCarta === "true";
-    }
-
-    setVisaoCarta(visaoCarta) {
-        this.cookieService.set("visaoCarta", visaoCarta);
-    }
-
     async criarRodada(jogadoresParticipantes, jogoId, rodadaNro) {
         var count = 0;
